referral_system/requirements.txt
asgiref==3.8.1
datadispatch==1.0.0
Django==4.2.7
django-environ==0.11.2
django-phonenumber-field==8.0.0
djangorestframework==3.14.0
drf-yasg==1.21.7
environs==11.0.0
inflection==0.5.1
marshmallow==3.21.3
packaging==24.2
phonenumbers==8.13.43
pillow==10.4.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
pytz==2024.2
PyYAML==6.0.2
setuptools==75.6.0
sqlparse==0.5.2
tasks==2.8.0
tzdata==2024.2
uritemplate==4.1.1

